Feature 314 expand mesage metadata

+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/services/OrganizationService.java
@@ -1,15 +1,19 @@
 package dev.ctrlspace.gendox.gendoxcoreapi.services;
 
 import dev.ctrlspace.gendox.gendoxcoreapi.exceptions.GendoxException;
-import dev.ctrlspace.gendox.gendoxcoreapi.model.Organization;
-import dev.ctrlspace.gendox.gendoxcoreapi.model.UserOrganization;
+import dev.ctrlspace.gendox.gendoxcoreapi.model.*;
 import dev.ctrlspace.gendox.gendoxcoreapi.model.authentication.UserProfile;
+import dev.ctrlspace.gendox.gendoxcoreapi.model.dtos.OrganizationDidDTO;
+import dev.ctrlspace.gendox.gendoxcoreapi.model.dtos.WalletKeyDTO;
 import dev.ctrlspace.gendox.gendoxcoreapi.model.dtos.criteria.OrganizationCriteria;
+import dev.ctrlspace.gendox.gendoxcoreapi.model.dtos.criteria.ProjectCriteria;
 import dev.ctrlspace.gendox.gendoxcoreapi.repositories.OrganizationRepository;
 import dev.ctrlspace.gendox.gendoxcoreapi.repositories.UserOrganizationRepository;
+import dev.ctrlspace.gendox.gendoxcoreapi.repositories.WalletKeyRepository;
 import dev.ctrlspace.gendox.gendoxcoreapi.repositories.specifications.OrganizationPredicates;
 import dev.ctrlspace.gendox.gendoxcoreapi.utils.constants.OrganizationRolesConstants;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -18,7 +22,10 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 @Service
@@ -29,14 +36,35 @@ public class OrganizationService {
     private OrganizationRepository organizationRepository;
     private UserOrganizationService userOrganizationService;
 
+    private OrganizationDidService organizationDidService;
+
+    private WalletKeyService walletKeyService;
+
+    private TypeService typeService;
+
+    private ProjectService projectService;
+
+    @Value("${walt-id.default-key.type}")
+    private String keyTypeName;
+    @Value("${walt-id.default-key.size}")
+    private Integer characterLength;
+
 
     @Autowired
     public OrganizationService(UserOrganizationRepository userOrganizationRepository,
                                OrganizationRepository organizationRepository,
-                               UserOrganizationService userOrganizationService) {
+                               UserOrganizationService userOrganizationService,
+                               OrganizationDidService organizationDidService,
+                               WalletKeyService walletKeyService,
+                               TypeService typeService,
+                               ProjectService projectService) {
         this.userOrganizationRepository = userOrganizationRepository;
         this.organizationRepository = organizationRepository;
         this.userOrganizationService = userOrganizationService;
+        this.organizationDidService = organizationDidService;
+        this.walletKeyService = walletKeyService;
+        this.typeService = typeService;
+        this.projectService = projectService;
 
     }
 
@@ -78,6 +106,23 @@ public class OrganizationService {
 
         userOrganizationService.createUserOrganization(ownerUserId, organization.getId(), OrganizationRolesConstants.ADMIN);
 
+        Type walletKeyType = typeService.getKeyTypeByName(keyTypeName);
+
+        WalletKeyDTO walletKeyDTO = WalletKeyDTO.builder()
+                .organizationId(organization.getId())
+                .keyType(walletKeyType)
+                .characterLength(characterLength)
+                .build();
+        WalletKey walletKey = walletKeyService.createWalletKey(walletKeyDTO);
+
+        OrganizationDid organizationDid = organizationDidService.createOrganizationDid(OrganizationDidDTO.builder()
+                        .createdAt(Instant.now())
+                        .updatedAt(Instant.now())
+                        .organizationId(organization.getId())
+                        .keyId(walletKey.getId())
+                .build(), "key");
+
+
         return organization;
     }
 
@@ -101,19 +146,66 @@ public class OrganizationService {
 
     }
 
-    public void deleteOrganization(UUID organizationId) throws Exception {
+    public void deactivateOrganization(UUID organizationId) throws GendoxException {
+
+        Organization organization = getById(organizationId);
+
+
+        if (("DEACTIVATED").equals(organization.getName())) {
+            return;
+        }
+
+        // Fetch all user organizations for this organization
         List<UserOrganization> userOrganizations = userOrganizationService.getUserOrganizationByOrganizationId(organizationId);
-        Organization organization = organizationRepository.findById(organizationId).orElse(null);
 
-        for(UserOrganization userOrganization: userOrganizations){
-            userOrganizationRepository.delete(userOrganization);
+        // Iterate through user organizations to handle both deletion and exception
+        for (UserOrganization userOrganization : userOrganizations) {
+            UUID userId = userOrganization.getUser().getId();
+
+            // Count the number of organizations the user is associated with
+            long count = userOrganizationRepository.countByUserId(userId);
+
+            if (count <= 1) {
+                // If the user has exactly one organization, throw an exception
+                throw new GendoxException("ORGANIZATION_DEACTIVATION_FAILED", "Cannot deactivate organization. User is associated with only one organization", HttpStatus.BAD_REQUEST);
+            } else {
+
+                deactivateAllOrgProjects(organizationId);
+                userOrganizationRepository.delete(userOrganization);
+            }
         }
 
-        if (organization != null) {
-            organizationRepository.deleteById(organizationId);
-        } else {
-            throw new GendoxException("ORGANIZATION_NOT_FOUND", "Organization not found with id: " + organizationId, HttpStatus.NOT_FOUND);
+        // Delete other associated data
+        organizationDidService.deleteOrganizationDidByOrganizationId(organizationId);
+        walletKeyService.deleteWalletKeyByOrganizationId(organizationId);
+
+        // Clear organization data and save the changes
+        clearOrgData(organization);
+        organizationRepository.save(organization);
+    }
+
+
+
+    private void deactivateAllOrgProjects(UUID organizationId) throws GendoxException {
+
+        ProjectCriteria criteria = new ProjectCriteria();
+        criteria.setOrganizationId(organizationId.toString());
+
+        Page<Project> projects = projectService.getAllProjects(criteria);
+
+        for (Project project : projects.getContent()) {
+            projectService.deactivateProject(project.getId());
         }
+    }
+
+    private void clearOrgData(Organization organization) {
+        organization.setName("DEACTIVATED");
+        organization.setDisplayName(null);
+        organization.setAddress(null);
+        organization.setPhone(null);
+        organization.setDeveloperEmail(null);
+        organization.setUpdatedAt(null);
+        organization.setCreatedAt(null);
     }
